pyproject.toml
[tool.poetry]
name = "pyswaggen"
version = "0.1.0"
description = "Generate python3 dataclasses from swagger definitions"
authors = ["Gerrit Jansen van Vuuren <gerritjvv@gmail.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.6"
dacite = "requests"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
